+++ packages/components/Affix/src/index.ts
@@ -1,0 +1,2 @@
+export { default as Affix } from './Affix.vue'
+export type * from './types'