+++ packages/theme/src/shortcuts/Affix.ts
@@ -1,0 +1,3 @@
+export const Affix = {
+  'lt-affix': 'sticky'
+}